+++ src/lib.rs
@@ -6,3 +6,8 @@ pub mod omikuji;
 pub mod options;
 pub mod timeline;
 pub mod yaml;
+<<<<<<< HEAD
+=======
+#[macro_use]
+extern crate horrorshow;
+>>>>>>> d91fd31392813c79a33cf5dc10eae06db2ce2613