+++ src/options/mod.rs
@@ -1,3 +1,4 @@
+pub mod htmlreport;
 pub mod level_tuning;
 pub mod profile;
-pub mod update_rules;
+pub mod update;